% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInput.R
\name{createInput}
\alias{createInput}
\title{Create Input for my package}
\usage{
createInput(fn)
}
\arguments{
\item{fn}{Name of my file to be transformed}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
