% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Example Data}
\format{
A data frame with 10 rows and 1 variables:
\describe{
\item{\code{V1}}{character quadruples}
}
}
\usage{
exampleData
}
\description{
List of 10 quadruples
}
\details{
Simple list of quadruples
}
\keyword{datasets}
