% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNRC.R
\name{findNRC}
\alias{findNRC}
\title{Find no-rainbow-coloring for input quadruples}
\usage{
findNRC(data, verbose = F)
}
\arguments{
\item{data}{The input data is the list created by \emph{createInput}, both the data table \emph{taxa} and \emph{data} is used}

\item{verbose}{Logical parameter if message should be printed; default F}
}
\value{
A list containing three sets:
\itemize{
\item the coloring per taxa
\item the colored quadruples
\item the result: \emph{fail} indicates decisiveness (no coloring was a no-rainbow 4-coloring), while \emph{NRC} indicates the first no-rainbow 4-coloring and the reason to call the set not-decisive
}
}
\description{
This function searches for a no-rainbow 4-coloring of the input quadruples. If there is one, then it represents one 4-way partition that is not covered by the input quadruples - hence the qaudruple set is not phylogenetic decisive.
}
\details{
The algorithm was taken from Ghazaleh Parvini
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[data.table]{copy}}
}
