% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNRC_helpeR_colorQuadFunc.R
\name{findNRC_helpeR_colorQuadFunc}
\alias{findNRC_helpeR_colorQuadFunc}
\title{Coloring of quadruples}
\usage{
findNRC_helpeR_colorQuadFunc(input, colors)
}
\arguments{
\item{input}{input data.table containing the columns taxa1, taxa2, taxa3, taxa4 and quadruple (first five columns of the data.table \emph{data} create by \emph{createInput()}, filtered for status == "input")}

\item{colors}{data.table containing all taxa with their original coding (e.g. a --> 1) and a column for the color. The four used colors are "green", "red", "blue", and "yellow". If no color is yet applied, all entries should be "nocol".}
}
\value{
The function returns an updated input data.table. In addition the the four numeric columns, and the quadruple column, there are now the colored taxa included, plus two columns with the sums per quadruple.
}
\description{
In this functions, the given input quadruples are colored with respect to a given coloring input. It also sums the number of colored quadruples (are there any uncolored taxa in any quadruple left?) and the sum of unique colors per quadruple (are there any quadruples with more than 3 colors?). It is a helper function for the \emph{finNRC} algorithm.
}
\details{
needs \emph{findNRC_helpeR_getColTaxaPerQuad} and \emph{findNRC_helpeR_getUniColTaxaPerQuad}
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[data.table]{copy}}
}
