% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNRC_helpeR_getColTaxaPerQuad.R
\name{findNRC_helpeR_getColTaxaPerQuad}
\alias{findNRC_helpeR_getColTaxaPerQuad}
\title{Get number of colored taxa per quadruple}
\usage{
findNRC_helpeR_getColTaxaPerQuad(x)
}
\arguments{
\item{x}{vector with four character elements, either "green", "red", "blue", "yellow" or "nocol".}
}
\value{
Sum of colored taxa for the input quadruple
}
\description{
Helper function for \emph{findNRC_helpeR_colorQuadFunc} to sum up all entries per row that are colred
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
