% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNRC_helpeR_getUniColTaxaPerQuad.R
\name{findNRC_helpeR_getUniColTaxaPerQuad}
\alias{findNRC_helpeR_getUniColTaxaPerQuad}
\title{Get number of unique colors per quadruple}
\usage{
findNRC_helpeR_getUniColTaxaPerQuad(x)
}
\arguments{
\item{x}{vector with four character elements, either "green", "red", "blue", "yellow" or "nocol".}
}
\value{
character containing all colors, seperated by "|"
}
\description{
Helper function for \emph{findNRC_helpeR_colorQuadFunc} to sum up all unique colors of a quadruple
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
