% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialCheck.R
\name{initialCheck}
\alias{initialCheck}
\title{Perform some initial checks}
\usage{
initialCheck(data, verbose = F)
}
\arguments{
\item{data}{Input data created by createInput()}

\item{verbose}{Logical parameter if message should be printed; default F}
}
\value{
Results of the three checks (OK or not OK)
}
\description{
Three checks, according to my master thesis
}
\details{
See master thesis, Lemma 1, Theorem 5 & 6
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[data.table]{as.data.table}}
}
