% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myAlgorithm.R
\name{myAlgorithm}
\alias{myAlgorithm}
\title{FUNCTION_TITLE}
\usage{
myAlgorithm(c)
}
\arguments{
\item{c}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
