% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAlgorithm.R
\name{runAlgorithm}
\alias{runAlgorithm}
\title{FUNCTION_TITLE}
\usage{
runAlgorithm(data)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
